remove unncessary method call and rename method name
